move function makeUserLogin to DuskTestCase.php
